Clean up minor spelling / grammar in Getting Started page
Hello, I work for Google but am currently contributing this as an individual. I'm interested in learning Bazel on the side, and as I'm going through the docs, I thought I'd submit code reviews for misc. Bazel site cleanup that I came across (especially for guides that act as a first impression for new users).

Let me know if you'd prefer I submit this patch in a different way. Thanks!

Closes #1529.

--
Reviewed-on: https://github.com/bazelbuild/bazel/pull/1529
MOS_MIGRATED_REVID=128035787