Shorten release baking period to 1 week (#180)
The original policy was created in times when [incompatible changes policy](https://docs.bazel.build/versions/master/backward-compatibility.html) didn't exist. With the policy in place, and Bazel Green team monitoring [the downstream pipeline](https://buildkite.com/bazel/bazel-at-head-plus-downstream) on the Bazel CI, and continuously adding more projects to the downstream pipeline, I believe we discover fewer regressions in Bazel release candidates then before.

On the other hand, we could use more time for users who want to test incompatible flags that are in their migration window and give feedback to the flag owner. Currently it happens that there are just few days between a release candidate is released and before next one is cut.

This change is an experiment to see if shorter baking period will result in faster releases, more time to give feedback on incompatible changes, while not regressing on the number of patch releases.