Deduplicate the release policy in favor of the playbook
Two avoid having two sources describing our release policy, remove
the one on the website in favor of a reference to our public release
play book.

Note: by this change we're dropping the requirement "A release can
only be released on a day where the next day is a business day."
which has never been an issue for an open-source release anyway.