Clarify how to write relnotes (#182)
* Clarify how to write relnotes

The automatic release notes generator does take into account the [INC] tag, so it should be used when the flag is actually flipped. Additionally, it is really helpful to the release manager if the RELNOTES themselves contain the name of the flag, what happened with the flag (added vs flipped), and the link to the github issue directly in the RELNOTES (as opposed to just in the commit description). Otherwise the RM has to search for keywords to find the actual github issue.

* remove link to relnotes.html

Co-authored-by: Danna Kelmer <dannark@dannark-macbookpro3.roam.corp.google.com>