Remove FAQ reference to "basil" (the herb)
The FAQ entry "How do you pronounce 'Bazel'" says it's pronounced the same as "basil" (the herb), and then gives the "BAY-el" pronunciation.

This makes things hopelessly confusing, as basil can be correctly pronounced _either_ to rhyme with "hazel" or to rhyme with "frazzle". In fact, many dictionaries list the "rhymes with frazzle" pronunciation first, with "rhymes with hazel" as a secondary pronunciation:

https://www.dictionary.com/browse/basil

https://www.merriam-webster.com/dictionary/basil

So I removed the "basil" reference entirely to avoid this confusion.